boldly v0.1.0 Boldly.ConversationInfo

Functions

change_conversation(conv)

create_conversation(attrs \\ %{})

Create a Conversation between a Brand and Creator.

Examples

iex> create_conversation(%{field: value}) {:ok, %Conversaton{}}

iex> create_conversation(%{field: bad_value}) {:error, %Ecto.Changeset{}}

delete_conversation(conv)

get_brand_conversations(brand_id)

Gets all conversations that the Brand is a part of.

Example

iex> get_brand_conversations(123) [%Conversation{}, ...]

get_conversation!(creator_id, brand_id)

Gets a single conversation.

Raises Ecto.NoResultsError if the Conversation does not exist.

Examples

iex> get_conversation!(123, 123) %Conversation{}

iex> get_conversation!(456, 456) ** (Ecto.NoResultsError)

get_creator_conversations(creator_id)

Gets all conversations that the Creator is a part of.

Example

iex> get_creator_conversations(123) [%Conversation{}, ...]

list_conversations()

Returns the list of all conversations.

Examples

iex> list_conversations() [%Conversation{}, ...]

update_conversation(conv, attrs \\ %{})