boldly v0.1.0 BoldlyWeb.BrandController

Functions

create(conn, map)

Creates a Brand account.

The attributes for the desired brand should be wrapped in a brand key at the top level of the JSON load.

Input fields are:

ecommerce: boolean, default = false
email: string
industries: string
location: string
values: string
password: string
name: string

Output fields can be seen in BoldlyWeb.BrandView.render/2

delete(conn, map)

Deletes a brand using the passed in id value. Returns an empty response on success

index(conn, params)

Lists all Brand accounts.

Returns a JSON list of all brands. Output fields can be seen in BoldlyWeb.BrandView.render/2.

show(conn, map)

Returns the attributes of the desired brand if it exists.

Output fields can be seen in BoldlyWeb.BrandView.render/2.

sign_in(conn, map)

Ensures that the login information is correct.

Should be passed the Email and Password under email and password keys in JSON.

Output fields can be seen in BoldlyWeb.BrandView.render/2

update(conn, map)

Updates the desired Brand's attributes.

At the top level of the JSON, there should be an id key which is the id of the Brand, and a brand key within which are the attributes that are being updated.

Returns the Brand with updated fields if successful, or an error if not.

Output fields can be seen in BoldlyWeb.BrandView.render/2.