boldly v0.1.0 BoldlyWeb.CampaignController

Functions

create(conn, map)

Creates a Campaign given the attributes are valid.

The campaign attributes should be wrapped in a campaign field at the top level of the JSON.

The fields should be:

launched_by: UUID of the Brand which is launching this Campaign
age_range: string
compensation: string
creator_responsibilities: string
description: string
desired_engagement_rate: integer
end_date: date
industry: string
interests: string
is_draft: boolean, default: false
location: string
name: string
perks: string
photo_reference: string
specific_to_location: boolean, default: false
start_date: date
values: string

Output fields can be seen in BoldlyWeb.CampaignView.render/2.

delete(conn, map)

Deletes Campaign with the given id.

Returns an empty response if successful.

get_camps_and_parts(conn, map)

get_pictures(campaigns)

index(conn, params)

Lists all Campaigns.

Output is a list of brands containing the fields listed in BoldlyWeb.CampaignView.render/2.

match_all_campaigns(conn, params)

show(conn, map)

Shows the Campaign with the matching id value.

Output fields can be seen in BoldlyWeb.CampaignView.render/2.

update(conn, map)

Updates the Campaign with id given under id and campaign attributes given under campaign keys.

Outputs newly updated object.

Output fields can be seen in BoldlyWeb.CampaignView.render/2.