boldly v0.1.0 BoldlyWeb.ConversationController

Functions

create(conn, map)

Creates a Conversation between a brand and a creator. The attributes for the desired conversation should be wrapped in conversation key at the top level of the JSON load.

Input fields are:

brand_id: id,
creator_id: id

Output fields can be seen in BoldlyWeb.ConversationView.render/2.

delete(conn, map)

Not yet implemented.

get_conversations(conn, map)

Returns all Conversations that the Creator or brand is a part of.

The payload for Creators should be {creator_id: id}, and for Brands should be {brand_id: id}.

Output fields can be seen in BoldlyWeb.ConversationView.render/2.

index(conn, params)

Lists all Conversations.

Returns a JSON list of all conversations. Output fields can be seen in BoldlyWeb.ConversationView.render/2.

show_conv(conn, map)

Returns the Conversation information of the brand and creator's conversation.

JSON payload should have, at the top level:

brand_id: id,
creator_id: id

Output fields can be seen in BoldlyWeb.ConversationView.render/2.

update(conn, map)

Not yet implemented.