boldly v0.1.0 BoldlyWeb.CreatorController

Functions

create(conn, map)

Creates a Creator account given valid attributes.

The attributes for the desired creator should be wrapped in a creator key at the top level of the JSON load.

Input fields are:

birthday: date
email: string
industry: string
interests: string
location: string
name: string
values: string
password: string

Output fields can be seen in BoldlyWeb.CreatorView.render/2.

delete(conn, map)

Deletes a Creator account if it exists.

Returns an empty respnse.

get_pictures(creators)

increment_views(conn, map)

index(conn, params)

Lists all Creator accounts.

Returns a JSON list of all creators. Output fields can be seen in BoldlyWeb.CreatorView.render/2.

show(conn, map)

Shows the specified creator if it exists.

Output fields can be seen in BoldlyWeb.CreatorView.render/2.

sign_in(conn, map)

Ensures that the email and password of the account are valid.

If the account doesn't exist, sends 401 and sends error message.

Output fields can be seen in BoldlyWeb.CreatorView.render/2.

update(conn, map)

Updates the Creator account with the desired attributes. The creator account's id should be at the top level of the JSON, and the updated attributes should be under a creator key at the top level of the JSON.

Output fields can be seen in BoldlyWeb.CreatorView.render/2.

update_engagement(conn, map)