boldly v0.1.0 BoldlyWeb.MessageController

Functions

create(conn, map)

Creates a Message.

The attributes for the desired Message should be wrapped in a message key at the top level of the JSON load.

Input fields are:

content: string,
conversation_id: id,
date: date,
sent_by_creator: boolean

Output fields can be seen in BoldlyWeb.MessageView.render/2.

delete(conn, map)

index(conn, params)

Lists all Messages.

Returns a JSON list of all Messages.

Output fields can be seen in BoldlyWeb.MessageView.render/2.

show(conn, map)

Returns the attributes of the desired Message if it exists.

Output fields can be seen in BoldlyWeb.MessageView.render/2.

show_conv(conn, map)

Returns all Messages from the Conversation.

Output fields can be seen in BoldlyWeb.MessageView.render/2.