boldly v0.1.0 BoldlyWeb.ParticipantController

Functions

activate_creator(conn, map)

apply_to_campaign(conn, map)

create(conn, map)

Creates a Participant for a Campaign given valid creation attributes.

Fields should be wrapped in a participant field at the top level of the JSON payload.

Fields should be:

campaign_uuid: uuid of Campaign of which the creator is participating
creator_uuid: The uuid of the Creator participating in the Campaign
is_active: boolean, default: false
is_deleted: boolean, default: false

Output fields can be seen in BoldlyWeb.ParticipantView.render/2.

deactivate_creator(conn, map)

delete(conn, map)

Deletes a Campaign Participant given a valid id under and id key.

Sends an empty response.

get_applicants_to_campaign(conn, map)

get_creators_in_campaign(conn, map)

index(conn, params)

Lists all campaign participants across all Campaigns.

Output fields can be seen in BoldlyWeb.ParticipantView.render/2.

match_creators(conn, map)

show(conn, map)

Displays the Campaign Participant info given a valid id.

Output fields can be seen in BoldlyWeb.ParticipantView.render/2.

update(conn, map)

Updates the Participant given a valid id under and id attribute and valid update attributes under a participant key.

Output fields can be seen in BoldlyWeb.ParticipantView.render/2.