boldly v0.1.0 BoldlyWeb.Router.Helpers

Module with named helpers generated from BoldlyWeb.Router.

Functions

brand_path(conn_or_endpoint, action)

brand_path(conn_or_endpoint, action, params)

brand_path(conn_or_endpoint, action, id, params)

brand_url(conn_or_endpoint, action)

brand_url(conn_or_endpoint, action, params)

brand_url(conn_or_endpoint, action, id, params)

campaign_path(conn_or_endpoint, action)

campaign_path(conn_or_endpoint, action, params)

campaign_path(conn_or_endpoint, action, id, params)

campaign_url(conn_or_endpoint, action)

campaign_url(conn_or_endpoint, action, params)

campaign_url(conn_or_endpoint, action, id, params)

contract_path(conn_or_endpoint, action)

contract_path(conn_or_endpoint, action, params)

contract_path(conn_or_endpoint, action, id, params)

contract_url(conn_or_endpoint, action)

contract_url(conn_or_endpoint, action, params)

contract_url(conn_or_endpoint, action, id, params)

creator_path(conn_or_endpoint, action)

creator_path(conn_or_endpoint, action, params)

creator_path(conn_or_endpoint, action, id, params)

creator_url(conn_or_endpoint, action)

creator_url(conn_or_endpoint, action, params)

creator_url(conn_or_endpoint, action, id, params)

participant_path(conn_or_endpoint, action)

participant_path(conn_or_endpoint, action, params)

participant_path(conn_or_endpoint, action, id, params)

participant_url(conn_or_endpoint, action)

participant_url(conn_or_endpoint, action, params)

participant_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

template_path(conn_or_endpoint, action)

template_path(conn_or_endpoint, action, params)

template_path(conn_or_endpoint, action, id, params)

template_url(conn_or_endpoint, action)

template_url(conn_or_endpoint, action, params)

template_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)